implement cache on mongodb gets
